refactor code for search by keyword and filter author using modular function
