+++ helpers/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   getCurrentDate: require('./dateHelper.js').getCurrentDate,
-  sortData: require('./dataHelper.js').sortData
+  sortData: require('./dataHelper.js').sortData,
+  searchAndFilterData: require('./dataHelper.js').searchAndFilterData
 }
